Allow to transcribe only sections of the audio (#57)
* add argument sections to predict.whisper

* code cleanup

* adding example of sections

* add extra data check on subset.wav such that offsets/durations are not outside of audio range

* VAD on mono

* Align the timestamps based on the removed audio segments again

* play around with 1 millisecond

* more examples on multiple-offsets

* add NEWS item, bump version

* unit test for multiple sections return the part where it has a voice instead of the non-voice and put thses skipped parts at the top

* CI

* put alignment timestamps of transcriptions with skipped segments in separate function

* add utils

* docs

* CI

* docs

* dev code for token_timestamps

* CI

* docs

* CI

* use format instead of timezone

* parse the timestamps with milliseconds

* make sure to parse out the milliseconds + test with token_timestamp

* rename sentences to voiced for later readability

* bump dependency version of data.table to when data.table contains nafill

* there is no guarantee on short segments that whisper.cpp adheres to the requested timeframe - it provides the whole segment - disable that unit test

* docs

* CI

* add segment_offset to keep track of within which offset the segment was part of

* add offset/duration as arguments in predict.whisper, include new column called segment_offset in output by default, make sure examples on stereo are run in language es

* increase duration to make sure we have a segment

* logs

* more liberal unit test

* more liberal unit test

+++ inst/tinytest/test-multiple-offsets.R
@@ -1,35 +1,53 @@
 library(audio.whisper)
 
 if(Sys.getenv("TINYTEST_CI", unset = "yes") == "yes"){
-  ## JFK example full fragment using tiny model
-  model <- whisper("tiny")
-  trans <- predict(model, newdata = system.file(package = "audio.whisper", "samples", "jfk.wav"), language = "en", 
-                   offset = c(0, 4000), duration = c(1*1500, 1*5000))
-  expect_equal(trans$n_segments, 2)
-  expect_equal(nrow(trans$data), 2)
-  if(file.exists(model$file)) file.remove(model$file)
-  
   ## Longer file
   download.file("https://github.com/jwijffels/example/raw/main/example.wav", "example.wav")
   model <- whisper("tiny")
   trans <- predict(model, newdata = "example.wav", language = "en", 
-                   offset = c(7*1000, 60*1000), duration = c(4*1500, 2*1000))
-  expect_equal(trans$n_segments, 2)
-  expect_equal(nrow(trans$data), 2)
+                   offset = c(7*1000, 60*1000), duration = c(6*1000, 5*1000))
+  expect_true(length(unique(trans$data$segment_offset)) > 1)
+  
+  ## Multiple sections
+  if(require(data.table) && require(audio)){
+  sections <- data.frame(start = c(7*1000, 60*1000), duration = c(6*1000, 5*1000))
+  trans    <- predict(model, newdata = "example.wav", language = "en", sections = sections)
+  expect_true("segment_offset" %in% colnames(trans$data))
   if(file.exists(model$file)) file.remove(model$file)
   if(file.exists(trans$params$audio)) file.remove(trans$params$audio)
+  }
   
   if(FALSE){
     library(audio.whisper)
     library(audio.vadwebrtc)
-    audio <- system.file(package = "audio.whisper", "samples", "stereo.wav")
+    library(av)
+    audio      <- system.file(package = "audio.whisper", "samples", "stereo.wav")
+    audio_mono <- audio
+    audio_mono <- tempfile(fileext = ".wav")
+    av_audio_convert(audio, channels = 1, output = audio_mono, format = "wav")
     ## Voice activity detection
-    vad    <- VAD(audio)
-    voiced <- is.voiced(vad, units = "milliseconds", silence_min = 1000)
+    vad    <- VAD(audio_mono)
+    voiced <- is.voiced(vad, units = "milliseconds", silence_min = 250)
     voiced <- subset(voiced, has_voice == TRUE)
+    voiced
+    p <- audio.whisper:::subset.wav(audio, offset = voiced$start, duration = voiced$duration)
+    p
+    file.copy(p$file, to = "onlyvoiced.wav", overwrite = TRUE)
     ## Transcription of voiced segments
-    path  <- system.file(package = "audio.whisper", "repo", "ggml-tiny.en-q5_1.bin")
-    model <- whisper(path)
-    trans <- predict(model, newdata = audio, language = "auto", offset = voiced$start, duration = voiced$duration, language = "es", trace = TRUE)
+    model <- whisper("tiny")
+    trans <- predict(model, newdata = audio, language = "es", sections = voiced)
+    trans <- predict(model, newdata = audio, language = "es", offset = voiced$start, duration = voiced$duration)
+    
+    audio <- "example.wav"
+    download.file("https://github.com/jwijffels/example/raw/main/example.wav", audio)
+    model <- whisper("tiny")
+    sections <- data.frame(start = c(7*1000, 60*1000), duration = c(6*1000, 2*1000))
+    trans <- predict(model, newdata = audio, language = "en", offset = sections$start, duration = sections$duration)
+    trans <- predict(model, newdata = audio, language = "en", sections = sections)
+    p <- audio.whisper:::subset.wav(audio, offset = sections$start, duration = sections$duration)
+    p
+    file.copy(p$file, to = "onlyvoiced.wav", overwrite = TRUE)
+    trans <- predict(model, newdata = audio, language = "en", sections = sections, token_timestamps = TRUE)
+    
   }
 }
